remove unnecessary predeclarations and reorder header inclusions
