+++ implementation/statement_evaluations.h
@@ -1,5 +1,5 @@
 #pragma once
-#include"smart_environments.h"
+#include"environment_sets.h"
 
 namespace expr
 {
@@ -100,29 +100,6 @@ namespace expr
 			}
 		}
 
-		inline void perform_all(executable&& tar, stack& loc, environment& env, std::ostream& errors)
-		{
-			for (auto&& it : std::move(tar.statements))
-			{
-				perform(std::move(it), loc, env.variables,env.functions, env.garbage, errors);
-			}
-		}
-
-		inline stack environment::run(executable&& tar, std::ostream& errors)
-		{
-			stack loc;
-			perform_all(std::move(tar), loc, *this, errors);
-			return loc;
-		}
-
-		inline stack_elem environment::evaluate(expression&& tar, std::ostream& errors)
-		{
-			executable to_run;
-			std::move(tar).into_executable(to_run);
-			stack v = run(std::move(to_run), errors);
-			assert_with_generic_logic_error(v.stuff.size() == 1);
-			return std::move(*v.stuff.begin());
-		}
 
 	}
 }
